// src/api/types/UpdateRehabilitationExaminationFormRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateRehabilitationExaminationFormRequest = {
  /**
   * @type string
   */
  patientName: string;
  /**
   * @type string, date
   */
  dateOfBirth: string;
  /**
   * @type integer, int32
   */
  age: number;
  /**
   * @type string | undefined
   */
  gender?: string;
  /**
   * @type string | undefined
   */
  ethnicity?: string;
  /**
   * @type string | undefined
   */
  job?: string;
  /**
   * @type string | undefined
   */
  address?: string;
  /**
   * @type string | undefined
   */
  contactPerson?: string;
  /**
   * @type string | undefined
   */
  phoneNumber?: string;
  /**
   * @type string, date
   */
  examinationDate: string;
  /**
   * @type string | undefined
   */
  chiefComplain?: string;
  /**
   * @type string | undefined
   */
  historyOfPresentIllness?: string;
  /**
   * @type string | undefined
   */
  pastMedicalHistory?: string;
  /**
   * @type string | undefined
   */
  allergies?: string;
  /**
   * @type string | undefined
   */
  medicalHistory?: string;
  /**
   * @type string | undefined
   */
  surgicalHistory?: string;
  /**
   * @type string | undefined
   */
  medicationHistory?: string;
  /**
   * @type string | undefined
   */
  familyHistory?: string;
  /**
   * @type string | undefined
   */
  rehabilitationExamination?: string;
  /**
   * @type string | undefined
   */
  laboratoryTestingAndImaging?: string;
  /**
   * @type string | undefined
   */
  diagnosis?: string;
  /**
   * @type string | undefined
   */
  interventionObjectives?: string;
  /**
   * @type string | undefined
   */
  therapyCourse?: string;
  /**
   * @type string | undefined
   */
  pulse?: string;
  /**
   * @type string | undefined
   */
  temperature?: string;
  /**
   * @type string | undefined
   */
  bloodPressure?: string;
  /**
   * @type string | undefined
   */
  breathingRate?: string;
  /**
   * @type number | undefined
   */
  height?: number;
  /**
   * @type number | undefined
   */
  weight?: number;
  /**
   * @type number | undefined
   */
  bmi?: number;
  /**
   * @type string | undefined
   */
  rehabilitationMethods?: string;
};